Correction de bugs

+++ Controllers/StolonsAdministrationController.cs
@@ -134,10 +134,13 @@ namespace Stolons.Controllers
                             vm.Stolon.UseSympathizer = false;
                             break;
                     }
-                    //Je supprime l'ancien logo
-                    _environment.DeleteFile(StolonLogoStockagePath, vm.Stolon.LogoFileName);
-                    //J'upload le nouveau et récupére son nom
-                    vm.Stolon.LogoFileName = _environment.UploadBase64Image( uploadLogo, StolonLogoStockagePath);
+                    if(!String.IsNullOrWhiteSpace(uploadLogo))
+                    {
+                        //Je supprime l'ancien logo
+                        _environment.DeleteFile(StolonLogoStockagePath, vm.Stolon.LogoFileName);
+                        //J'upload le nouveau et récupére son nom
+                        vm.Stolon.LogoFileName = _environment.UploadBase64Image(uploadLogo, StolonLogoStockagePath);
+                    }
                     _context.Update(vm.Stolon);
                     await _context.SaveChangesAsync();
                 }
